Added space bar randomize
